Testing: improve test coverage
